backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 資料庫
sqlalchemy==1.4.47
databases[aiosqlite]==0.7.0
alembic==1.12.1

# 任務佇列
celery==5.3.6
redis==5.0.1

# YouTube 下載
yt-dlp==2023.11.16

# 驗證與安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 工具庫
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0  # 確保這一行存在

# 開發工具（可選）
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
flower>=2.0
